requirements.txt
Flask==3.0.3
openai==1.30.4
load-dotenv==0.1.0